Add schema evolution to s3.to_csv (#799)
